+++ include/core/parameters.h
@@ -1173,6 +1173,9 @@ namespace Parameters
     // Kinsol solver strategy
     KinsolStrategy kinsol_strategy;
 
+    // Normalize residual by triangulation volume
+    bool normalize_residual_by_volume;
+
     // Tolerance
     double tolerance;